Adjusted checkstyle rules
Removed the custom ordering for imports because intellij already does a different ordering by default which I think is just as valid.
Also removed the required indentation level for javadoc continuation because it was bugging and refused to to allow further comments that weren't continuations.
Raised the line length limit from 125 to 200 because it was too short and I had to break up comments after a few words which was very annoying.